+++ src/components/Footer.tsx
@@ -25,7 +25,6 @@ export const Footer: React.FC<Props> = ({
         {`${todosCounter} items left`}
       </span>
 
-      {/* Active link should have the 'selected' class */}
       <nav className="filter" data-cy="Filter">
         {filters.map(filter => (
           <a
@@ -44,7 +43,6 @@ export const Footer: React.FC<Props> = ({
         ))}
       </nav>
 
-      {/* this button should be disabled if there are no completed todos */}
       <button
         type="button"
         className="todoapp__clear-completed"
