+++ src/components/Header.tsx
@@ -38,7 +38,6 @@ export const Header: React.FC<Props> = ({
         onClick={completedAllTodos}
       />
 
-      {/* Add a todo on form submit */}
       <form
         onSubmit={event => {
           event.preventDefault();
