+++ src/components/TempTodo.tsx
@@ -22,7 +22,6 @@ export const TempTodo: React.FC<Props> = ({ todo }) => {
         {todo.title}
       </span>
 
-      {/* Remove button appears only on hover */}
       <button
         type="button"
         className="todo__remove"
